Bug fix: Padding and converting the bit addition
The padding was not correct. Regardless of the size of the string, it
will always be padded up to 448 over a multiple of 12, then add
padding/length.

Also there was an error in converting the addition of the two bit
strings. There needs to be a padding of zeros at the beginning to get
the number up to a 32 bit number.